working on ryan supplement (stash test)
i'm trying the stash feature

working on manuscript changes in the editTobit branch. I wanted to commit and merge my code-related changes

but I wasn't ready to commit my manuscript changes, especially not in that branch. So i wanted to stash the changes, commit the coding changes, merge the editTobit branch with the master, then pick up the changes I'd been making and just work on those in the master branch.