correct for multiple comparisons in paired tests
Conclusion of new analysis:
	Suwa: It's easy to argue that correct for multiple comparisons does not change which pairs are significantly different. In the clear.
	Tornio: Only 1 pair was different before, and no matter what correction is chosen, it is no longer significant. So the only way it changes is if no correction is made.

For the Tornio pair that was originally significant (AOD, the volcano thing), I think the right thing to do is to report the uncorrected probability of a type 1 error (~2.5%), as well as the corrected probability of a type 1 error (~15%). We do this to highlight that this is the pair that is most likely to be different before and after, while still being honest about the probability.

@lawinslow thoughts? (also, do I need to tag you in commits?)