ryan's edits to manuscript
Email I sent:

Hey Folks!

Attached are some edits and comments. I didn't get to the Supplement, and got less detailed towards the end (ran out of steam, and started thinking that a lot of the text might be rewritten).

I think we have a lot of work to do on setting up the motivation for our study.

I think it is important to start the paper with clear questions that we are uniquely positioned to answer b/c of our data set. We don't have clear questions, and we don't say why the uniqueness of our data is advantageous.

Currently, much of the manuscript reads in a manner tantamount to saying "we have a really awesome data set, we calculated statistics on those data, and here is some context that supports our results". I.e., we have really good results and discussion, but we don't have a great setup. We need questions that we answer with statistics, not statistics that we support with literature.

Do you all see a similar issue, or do you understand the point I'm making? I think we're definitely getting there and have most of the right parts, we just need a better lead in. With that lead, we'll have a great story to tell :)

Happy New Year!