addresses #22 by showing AIC profile across BP's
with and without the 1867 data point. The reviewer mentioned 1756, but I can confirm that the date of the breakpoint (1867) does not change with the removal or inclusion of 1756. 1867 seems like more of an "extreme" or outlier, so I proceeded with an analysis that answers the reviewer's question assuming that the reviewer had intended to write "1867" instead of "1756".